+++ CHANGELOG.md
@@ -1,3 +1,22 @@
+# <a href='https://github.com/mrjackwills/staticpi_vue/releases/tag/v0.1.11'>v0.1.11</a>
+### 2024-05-07
+
+### Chores
++ dependencies updated, [8501f1d0](https://github.com/mrjackwills/staticpi_vue/commit/8501f1d078c4b0ee30c291236211c1c18a0f44a7), [7957a9df](https://github.com/mrjackwills/staticpi_vue/commit/7957a9df51d9654d9f1dc117c7c39937b63693cf)
+
+### Features
++ debug timeout in axios service, [b7006b38](https://github.com/mrjackwills/staticpi_vue/commit/b7006b381ce3742a5de21b8cc40b69a0485991a7)
+
+### Fixes
++ key app bar links, [be952f1b](https://github.com/mrjackwills/staticpi_vue/commit/be952f1bafe5f42e05285f052f5f981c6c300365)
++ use window.setTimeout, [a3725e8f](https://github.com/mrjackwills/staticpi_vue/commit/a3725e8f444cd6d427244879684d8167a4114e61)
++ set host in vite.config.ts, [3e2ac9cf](https://github.com/mrjackwills/staticpi_vue/commit/3e2ac9cf055642a3528c5b4c02784e834e9fce60)
++ loading bar sticky on page top, [132cfb23](https://github.com/mrjackwills/staticpi_vue/commit/132cfb238b017e936ea44605b0daf86f8ed72df1)
++ change links key, [84352397](https://github.com/mrjackwills/staticpi_vue/commit/843523973865170f969c9ed36354b2a78db8f88a)
+
+### Refactors
++ dead code removed, [f0da5e30](https://github.com/mrjackwills/staticpi_vue/commit/f0da5e30aea6925086d905303399228f0c77251e)
+
 # <a href='https://github.com/mrjackwills/staticpi_vue/releases/tag/v0.1.10'>v0.1.10</a>
 ### 2024-04-01